+++ src/commands/botinfo.ts
@@ -3,15 +3,13 @@ import * as config from '../../config.json';
 import os from 'node:os';
 
 import type { Command } from "../types";
-import { scamDB, serverDB, startup } from "../bot";
+import { scamDB, serverDB, startup } from "..";
 import { revision } from '../helpers';
 
 
-const botInfoTemplate = (client: Client, guildCacheSize: number[]) => `
+const botInfoTemplate = (client: Client) => `
 Bot Name: '${client.user?.tag}'
-Guild Count: ${guildCacheSize.reduce((a, b) => a + b, 0)}
-Shard Count: ${client.shard?.count}
-Shard Latency: ${Math.round(client.ws.ping)}ms
+Guild Count: ${client.guilds.cache.size}
 Startup Time: <t:${Math.floor(
   startup.getTime() / 1000
 )}:D> <t:${Math.floor(
@@ -31,7 +29,7 @@ const systemInformationTemplate = (hostname: string) => `
 )}:R>
 			`;
 
-const embed = (client: Client, guildCacheSize: number[], hostname: string) => new EmbedBuilder()
+const embed = (client: Client, hostname: string) => new EmbedBuilder()
   .setTitle('Bot Information')
   .setTimestamp(new Date())
   .setFields([{
@@ -40,7 +38,7 @@ const embed = (client: Client, guildCacheSize: number[], hostname: string) => ne
   },
   {
     'name': 'Bot Info',
-    'value': botInfoTemplate(client, guildCacheSize)
+    'value': botInfoTemplate(client)
   }
   ])
   .setFooter({
@@ -51,22 +49,18 @@ export default {
   data: new SlashCommandBuilder().setName("botinfo").setDescription("Get information about the client."),
   executeInteraction: async function (interaction: ChatInputCommandInteraction<CacheType>) {
     const client = interaction.client;
-    const guildCacheSize = await client.shard?.fetchClientValues("guilds.cache.size") as number[] | undefined
-    if (guildCacheSize === undefined) throw new Error('Failed to fetch guild cache sizes');
     const hostname = config.owners.includes(interaction.user.id) === true ? os.hostname() : os.hostname().replace(/./g, '•');
 
     interaction.reply({
-      embeds: [embed(client, guildCacheSize, hostname)]
+      embeds: [embed(client, hostname)]
     });
   },
   executeMessage: async function (message: Message<boolean>) {
     const client = message.client;
-    const guildCacheSize = await client.shard?.fetchClientValues("guilds.cache.size") as number[] | undefined
-    if (guildCacheSize === undefined) throw new Error('Failed to fetch guild cache sizes');
     const hostname = config.owners.includes(message.author.id) === true ? os.hostname() : os.hostname().replace(/./g, '•');
 
     message.reply({
-      embeds: [embed(client, guildCacheSize, hostname)]
+      embeds: [embed(client, hostname)]
     });
   }
 } as Command;
