+++ src/index.ts
@@ -1,28 +1,93 @@
+export const startup = new Date();
+
 import * as Discord from 'discord.js';
 import * as config from '../config.json';
-import type { MessageData } from './types';
-import path from 'node:path';
+import { loadCommands, updateDatabase } from './helpers';
+import { checkAttachments, checkMessageContent } from './helpers/checkers';
+import WebSocket from 'ws';
+import type { Command, ScamWSData, serverDBData } from './types';
 
-const manager = new Discord.ShardingManager(path.join(__dirname, '.', 'bot.js'), {
-  token: config.discord.token
+export let scamDB: string[] = [];
+export let serverDB: serverDBData = [];
+export let lastIdPerGuild: {
+  messageID: string;
+  userID: string;
+  guildID: string;
+}[] = [];
+let commands: Command[] = [];
+
+const bot = new Discord.Client({
+  intents: [
+    Discord.GatewayIntentBits.Guilds,
+    Discord.GatewayIntentBits.GuildMessages,
+    Discord.GatewayIntentBits.GuildMessageReactions,
+    Discord.GatewayIntentBits.GuildBans,
+    Discord.GatewayIntentBits.GuildMembers,
+    Discord.GatewayIntentBits.MessageContent
+  ],
+  partials: [
+    Discord.Partials.GuildMember,
+    Discord.Partials.Message,
+    Discord.Partials.Channel,
+    Discord.Partials.User
+  ]
 });
 
-manager.on('shardCreate', shard => {
-  console.log(`Launched shard ${shard.id}`)
-  shard.on("ready", () => {
-    console.log(`[DEBUG/SHARD] Shard ${shard.id} connected to Discord's Gateway.`)
-    // Sending the data to the shard.
-    shard.send({
-      type: "activity",
-      data: {
-        status: config.discord.status as unknown as Discord.PresenceStatusData,
-        activities: [{
-          type: config.discord.status.activities[0].type as unknown as Discord.ActivityType,
-          name: config.discord.status.activities[0].name
-        }]
-      }
-    } as MessageData);
-  })
+const scamSocket = new WebSocket(config.scams.scamSocket, {
+  headers: {
+    "User-Agent": "ScamBaiter/1.0; Chris Chrome#9158",
+    "X-Identity": "ScamBaiter/1.0; Chris Chrome#9158",
+  }
 });
 
-manager.spawn(config.discord.shardConfig);
+scamSocket.onopen = () => {
+  console.log("Connected to scam socket.");
+};
+
+scamSocket.onmessage = (msg) => {
+  const data = JSON.parse(msg.data as string) as ScamWSData;
+  if (data.type === "add") {
+    // Get all the entries in "data.domains" array and push to db
+    scamDB.push(...data.domains);
+  } else if (data.type === "delete") {
+    // Get all the entries in "data.domains" array and remove from db
+    scamDB = scamDB.filter(item => !data.domains.includes(item));
+  }
+};
+
+bot.once('ready', async () => {
+  console.log(`Logged in as ${bot.user?.tag}!`);
+
+  commands = await loadCommands()
+  await updateDatabase(scamDB, serverDB);
+});
+
+bot.on('interactionCreate', async interaction => {
+  if (!interaction.isChatInputCommand()) return;
+
+  // Check if command exists
+  const command = commands.find(cmd => cmd.data.name === interaction.commandName);
+  if (!command) return;
+  command.executeInteraction(interaction);
+});
+
+bot.on('messageCreate', async message => {
+  if (message.author.id == bot.user?.id) return;
+
+  const prefix = "$";
+  const args = message.content.slice(prefix.length).trim().split(/ +/g);
+  const cmd = args.shift()?.toLowerCase();
+
+  if (cmd !== 'check') {
+    const contentCheckResult = await checkMessageContent(message);
+    const attachmentCheckResult = await checkAttachments(message);
+    if (contentCheckResult || attachmentCheckResult) return;
+  }
+
+  // Check if command exists
+  const command = commands.find(command => command.data.name === cmd);
+  if (!command) return;
+  command.executeMessage(message, args);
+});
+
+bot.login(config.discord.token);
